HTTP analysis: do not depend on webconnectivity.TestKeys
Instead use urlgetter.TestKeys and webconnectivity.ControlResult.

This way we don't put too much constraints on the structure of
the TestKeys and we can enable further change.

Part of https://github.com/ooni/probe-engine/issues/776
